+++ include/common.php
@@ -1,29 +1,31 @@
 <?php
 
 require dirname(__DIR__) . '/vendor/autoload.php';
-include dirname(__DIR__) . '/config.php';
 
 use App\Legacy\Entity\Session;
 use App\Legacy\Entity\User;
 use App\Legacy\MySQLiDatabase;
+use Symfony\Component\Dotenv\Dotenv;
+
+$dotenv = new Dotenv();
+$dotenv->loadEnv(dirname(__DIR__) . '/.env');
 
 if (
-    !isset($database_host) ||
-    !isset($database_instance) ||
-    !isset($database_password) ||
-    !isset($database_user) ||
-    !isset($session_max_lifetime) ||
-    !isset($todo_priority)
+    !isset($_ENV['DATABASE_HOST']) ||
+    !isset($_ENV['DATABASE_INSTANCE']) ||
+    !isset($_ENV['DATABASE_PASSWORD']) ||
+    !isset($_ENV['DATABASE_USER']) ||
+    !isset($_ENV['SESSION_MAX_LIFETIME'])
 ) {
     throw new \Exception('Missing configuration values');
 }
 
 $GLOBALS['db'] = new MySQLiDatabase();
-$GLOBALS['db']->connect($database_host, $database_user, $database_password, $database_instance);
+$GLOBALS['db']->connect($_ENV['DATABASE_HOST'], $_ENV['DATABASE_USER'], $_ENV['DATABASE_PASSWORD'], $_ENV['DATABASE_INSTANCE']);
 
 session_cache_limiter('nocache');
 
-$GLOBALS['session_handler'] = new Session($GLOBALS['db'], $session_max_lifetime);
+$GLOBALS['session_handler'] = new Session($GLOBALS['db'], $_ENV['SESSION_MAX_LIFETIME']);
 $GLOBALS['session_handler']->initialize();
 session_name('cwci'); // cookie in Welsh
 session_start();
@@ -45,6 +47,11 @@ if ($GLOBALS['user']->getTimezone() != '') {
     $query = 'SET time_zone="' . addslashes($GLOBALS['user']->getTimezone()) . '"';
     $GLOBALS['db']->query($query);
 }
+
+$todo_priority = [
+    'high' => 1,
+    'low'  => 10,
+];
 
 $todo_priority['normal'] = intval((($todo_priority['low'] - $todo_priority['high']) / 2) + $todo_priority['high']);
 $GLOBALS['todo_priority'] = $todo_priority;